out_stdout: handle after 2038 timestamps

+++ plugins/out_stdout/stdout.c
@@ -36,6 +36,7 @@
 #include <cprofiles/cprof_decode_msgpack.h>
 
 #include <msgpack.h>
+#include <inttypes.h>
 #include "stdout.h"
 
 
@@ -402,7 +403,7 @@ static void cb_stdout_flush(struct flb_event_chunk *event_chunk,
 
             printf("[%zd] %s: [[", cnt++, event_chunk->tag);
 
-            printf("%"PRId32".%09lu, ", (int32_t) log_event.timestamp.tm.tv_sec,
+            printf("%"PRId64".%09lu, ", (int64_t) log_event.timestamp.tm.tv_sec,
                     log_event.timestamp.tm.tv_nsec);
 
             msgpack_object_print(stdout, *log_event.metadata);
